+++ src/renderable/imagelayer.js
@@ -93,7 +93,7 @@ class ImageLayer extends Sprite {
         this.repeat = settings.repeat || "repeat";
 
         // on context lost, all previous textures are destroyed
-        event.on(event.WEBGL_ONCONTEXT_RESTORED, this.createPattern, this);
+        event.on(event.ONCONTEXT_RESTORED, this.createPattern, this);
     }
 
     /**
@@ -294,7 +294,7 @@ class ImageLayer extends Sprite {
     destroy() {
         pool.push(this.ratio);
         this.ratio = undefined;
-        event.off(event.WEBGL_ONCONTEXT_RESTORED, this.createPattern);
+        event.off(event.ONCONTEXT_RESTORED, this.createPattern);
         super.destroy();
     }