+++ src/video/webgl/glshader.js
@@ -284,7 +284,7 @@ class GLShader {
         this.uniforms = extractUniforms(this.gl, this);
 
         // destroy the shader on context lost (will be recreated on context restore)
-        event.on(event.WEBGL_ONCONTEXT_LOST, this.destroy, this);
+        event.on(event.ONCONTEXT_LOST, this.destroy, this);
     }
 
     /**